tests/wpt/web-platform-tests/html/semantics/forms/textfieldselection/selection-not-application-textarea.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>text field selection (textarea)</title>
<link rel="author" title="Denis Ah-Kang" href="mailto:denis@w3.org">
<link rel=help href="https://html.spec.whatwg.org/multipage/#textFieldSelection">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
  test(function() {
    var el = document.createElement("textarea");
    assert_equals(el.selectionStart, 0);
    assert_equals(el.selectionEnd, 0);
    el.selectionStart = 1;
    el.selectionEnd = 1;
    el.selectionDirection = "forward";
    el.setRangeText("foobar");
    el.setSelectionRange(0, 1);
  }, "text field selection for the input textarea");
</script>
